# docs/PRD/0004_API_계약/openapi3.json
{
  "components" : {
    "schemas" : {
...
      "AdminClassResponse" : {
        "properties" : {
          "bufferMin" : {
            "format" : "int32",
            "type" : "integer"
          },
          "category" : {
            "type" : "string"
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "durationMin" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "imageUrl" : {
            "type" : [ "string", "null" ]
          },
          "name" : {
            "type" : "string"
          },
          "passEligible" : {
            "type" : "boolean"
          },
          "preparationInfo" : {
            "type" : [ "string", "null" ]
          },
          "price" : {
            "format" : "int64",
            "type" : "integer"
          },
          "status" : {
            "enum" : [ "ACTIVE", "INACTIVE" ],
            "type" : "string"
          },
          "targetAudience" : {
            "type" : [ "string", "null" ]
          }
        },
        "required" : [ "bufferMin", "category", "description", "durationMin", "id", "imageUrl", "name", "passEligible", "preparationInfo", "price", "status", "targetAudience" ],
        "type" : "object"
      },
...
      "ChangePasswordRequest" : {
        "properties" : {
          "currentPassword" : {
            "maxLength" : 72,
            "minLength" : 8,
            "type" : "string"
          },
          "newPassword" : {
            "maxLength" : 72,
            "minLength" : 8,
            "type" : "string"
          }
        },
        "required" : [ "currentPassword", "newPassword" ],
        "type" : "object"
      },
...
      "CreateClassRequest" : {
        "properties" : {
          "bufferMin" : {
            "format" : "int32",
            "type" : "integer"
          },
          "category" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "durationMin" : {
            "format" : "int32",
            "type" : "integer"
          },
          "imageUrl" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "passEligible" : {
            "type" : "boolean"
          },
          "preparationInfo" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "price" : {
            "format" : "int64",
            "maximum" : 9007199254740991,
            "minimum" : 10,
            "type" : "integer"
          },
          "targetAudience" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "bufferMin", "category", "durationMin", "name", "passEligible", "price" ],
        "type" : "object"
      },
...
      "CustomerLoginRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "email", "password" ],
        "type" : "object"
      },
...
      "CustomerUserResponse" : {
        "properties" : {
          "email" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "localPasswordEnabled" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          },
          "phone" : {
            "type" : [ "string", "null" ]
          },
          "phoneVerified" : {
            "type" : "boolean"
          }
        },
        "required" : [ "email", "id", "localPasswordEnabled", "name", "phone", "phoneVerified" ],
        "type" : "object"
      },
...
      "PasswordReauthenticationRequest" : {
        "properties" : {
          "currentPassword" : {
            "maxLength" : 72,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "currentPassword" ],
        "type" : "object"
      },
...
      "PolicyAcceptanceRequest" : {
        "properties" : {
          "privacyAccepted" : {
            "type" : "boolean"
          },
          "privacyVersion" : {
            "minLength" : 1,
            "type" : "string"
          },
          "termsAccepted" : {
            "type" : "boolean"
          },
          "termsVersion" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "privacyAccepted", "privacyVersion", "termsAccepted", "termsVersion" ],
        "type" : "object"
      },
...
      "ProductResponse" : {
        "properties" : {
          "available" : {
            "type" : "boolean"
          },
          "careInstructions" : {
            "type" : [ "string", "null" ]
          },
          "category" : {
            "type" : [ "string", "null" ]
          },
          "description" : {
            "type" : [ "string", "null" ]
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "imageUrl" : {
            "type" : [ "string", "null" ]
          },
          "name" : {
            "type" : "string"
          },
          "price" : {
            "format" : "int64",
            "type" : "integer"
          },
          "productionLeadDays" : {
            "format" : "int32",
            "type" : [ "integer", "null" ]
          },
          "quantity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "specification" : {
            "type" : [ "string", "null" ]
          },
          "status" : {
            "enum" : [ "ACTIVE", "INACTIVE" ],
            "type" : "string"
          },
          "type" : {
            "enum" : [ "READY_STOCK", "MADE_TO_ORDER" ],
            "type" : "string"
          }
        },
        "required" : [ "available", "careInstructions", "category", "description", "id", "imageUrl", "name", "price", "productionLeadDays", "quantity", "specification", "status", "type" ],
        "type" : "object"
      },
...
      "ResetPasswordRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "newPassword" : {
            "maxLength" : 72,
            "minLength" : 8,
            "type" : "string"
          },
          "phone" : {
            "minLength" : 1,
            "type" : "string"
          },
          "verificationCode" : {
            "minLength" : 1,
            "pattern" : "^[0-9]{6}$",
            "type" : "string"
          }
        },
        "required" : [ "email", "newPassword", "phone", "verificationCode" ],
        "type" : "object"
      },
...
      "SignupRequest" : {
        "properties" : {
          "email" : {
            "format" : "email",
            "minLength" : 1,
            "type" : "string"
          },
          "name" : {
            "minLength" : 1,
            "type" : "string"
          },
          "password" : {
            "maxLength" : 72,
            "minLength" : 8,
            "type" : "string"
          },
          "phone" : {
            "minLength" : 1,
            "type" : "string"
          },
          "policyAcceptance" : {
            "$ref" : "#/components/schemas/PolicyAcceptanceRequest"
          },
          "verificationCode" : {
            "minLength" : 1,
            "pattern" : "^[0-9]{6}$",
            "type" : "string"
          }
        },
        "required" : [ "email", "name", "password", "phone", "policyAcceptance", "verificationCode" ],
        "type" : "object"
      },
...
      "UpdateClassRequest" : {
        "properties" : {
          "category" : {
            "maxLength" : 30,
            "minLength" : 0,
            "type" : "string"
          },
          "description" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "imageUrl" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          },
          "name" : {
            "maxLength" : 100,
            "minLength" : 0,
            "type" : "string"
          },
          "passEligible" : {
            "type" : "boolean"
          },
          "preparationInfo" : {
            "maxLength" : 2000,
            "minLength" : 0,
            "type" : "string"
          },
          "price" : {
            "format" : "int64",
            "maximum" : 9007199254740991,
            "minimum" : 10,
            "type" : "integer"
          },
          "targetAudience" : {
            "maxLength" : 1000,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [ "category", "name", "passEligible", "price" ],
        "type" : "object"
      },
      "UpdateClassStatusRequest" : {
        "properties" : {
          "status" : {
            "enum" : [ "ACTIVE", "INACTIVE" ],
            "type" : "string"
          }
        },
        "required" : [ "status" ],
        "type" : "object"
      },
...
      "UpdateProductStatusRequest" : {
        "properties" : {
          "status" : {
            "enum" : [ "ACTIVE", "INACTIVE" ],
            "type" : "string"
          }
        },
        "required" : [ "status" ],
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "title" : "happyGallery API",
    "version" : "v1"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/classes" : {
      "get" : {
        "operationId" : "listClasses",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/AdminClassResponse"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-class-controller" ]
      },
      "post" : {
        "operationId" : "createClass",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateClassRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminClassResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "admin-class-controller" ]
      }
    },
    "/api/v1/admin/classes/{id}" : {
      "patch" : {
        "operationId" : "updateClass",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateClassRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminClassResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-class-controller" ]
      }
    },
    "/api/v1/admin/classes/{id}/status" : {
      "patch" : {
        "operationId" : "changeAdminClassStatus",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateClassStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminClassResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-class-controller" ]
      }
    },
...
    "/api/v1/admin/products/{id}/status" : {
      "patch" : {
        "operationId" : "changeStatus",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProductStatusRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProductResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-product-controller" ]
      }
    },
...
    "/api/v1/auth/login" : {
      "post" : {
        "operationId" : "loginCustomer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CustomerLoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerUserResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "customer-auth-controller" ]
      }
    },
...
    "/api/v1/auth/password/reset" : {
      "post" : {
        "operationId" : "resetCustomerPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "customer-credential-controller" ]
      }
    },
    "/api/v1/auth/signup" : {
      "post" : {
        "operationId" : "signupCustomer",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SignupRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CustomerUserResponse"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "customer-auth-controller" ]
      }
    },
...
    "/api/v1/me/password" : {
      "patch" : {
        "operationId" : "changeMyPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ChangePasswordRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "customer-credential-controller" ]
      }
    },
...
    "/api/v1/me/reauthentication/password" : {
      "post" : {
        "operationId" : "reauthenticateMyPassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordReauthenticationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        },
        "tags" : [ "customer-credential-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "description" : "Same-origin API",
    "url" : "/"
  } ]
}